Add interfaces for each service, inject interface instead of implementation to invert dependencies
