+++ src/main/java/crs/projects/mockbank/filter/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package crs.projects.mockbank.filter;
 
-import crs.projects.mockbank.config.JwtService;
+import crs.projects.mockbank.service.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -15,7 +15,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-
 import java.io.IOException;
 
 @Component
