+++ src/main/java/crs/projects/mockbank/service/JwtService.java
@@ -1,0 +1,20 @@
+package crs.projects.mockbank.service;
+
+import io.jsonwebtoken.Claims;
+import org.springframework.security.core.userdetails.UserDetails;
+import java.util.Map;
+import java.util.function.Function;
+
+public interface JwtService {
+
+    String extractUsername(String token);
+
+    String generateToken(UserDetails userDetails);
+
+    String generateToken(Map<String, Object> extraClaims, UserDetails userDetails);
+
+    boolean isTokenValid(String token, UserDetails userDetails);
+
+    <T> T extractClaim(String token, Function<Claims, T> claimsResolver);
+
+}
